Control multiple outlets
Subscribe to a single feed, "bl", and use it to control outlets #1 and
#1. At this point, I'm pausing the project. The range of the RF
transmitter isn't enough to control the lights in my bedroom from the
upstairs floor where the router is. This means I'd have to get a WiFi
enabled Arduino to put the transmitter in my bedroom. But, I've seen how
unreliable the unit is when connected via Ethernet, so I'm not convinced
an $80 WiFi board will work reliabily. Also, while it doesn't drop
packets anymore, it still fails to respond to commands sometimes (over
ethernet). It might be time to look into using the Raspberry Pi.